+++ include/appli.h
@@ -1,0 +1,16 @@
+#include <stdio.h>
+
+/**
+ *\brief lire le choix saisi par l'utilisateur lors d l'affichage du menu principal de l'application
+ *\return: int 
+*/
+
+int lire_choix();
+
+/**
+ *\brief affichager le menu principal de l'application
+*/
+
+void print_menu();
+
+
