--- src/include/appli.h
@@ -1,7 +1,0 @@
-#include <stdio.h>
-
-/**
- * Afficher le menu
- *
- */
-void print_menu();
